+++ packages/opentelemetry-exporter-collector/src/CollectorTraceExporterBase.ts
